Rename schema validotors directory to match Ruby namespace
