+++ lib/openapi_parser/schema_validator.rb
@@ -1,18 +1,18 @@
-require_relative 'schema_validators/options'
-require_relative 'schema_validators/enumable'
-require_relative 'schema_validators/minimum_maximum'
-require_relative 'schema_validators/base'
-require_relative 'schema_validators/string_validator'
-require_relative 'schema_validators/integer_validator'
-require_relative 'schema_validators/float_validator'
-require_relative 'schema_validators/boolean_validator'
-require_relative 'schema_validators/object_validator'
-require_relative 'schema_validators/array_validator'
-require_relative 'schema_validators/any_of_validator'
-require_relative 'schema_validators/all_of_validator'
-require_relative 'schema_validators/one_of_validator'
-require_relative 'schema_validators/nil_validator'
-require_relative 'schema_validators/unspecified_type_validator'
+require_relative 'schema_validator/options'
+require_relative 'schema_validator/enumable'
+require_relative 'schema_validator/minimum_maximum'
+require_relative 'schema_validator/base'
+require_relative 'schema_validator/string_validator'
+require_relative 'schema_validator/integer_validator'
+require_relative 'schema_validator/float_validator'
+require_relative 'schema_validator/boolean_validator'
+require_relative 'schema_validator/object_validator'
+require_relative 'schema_validator/array_validator'
+require_relative 'schema_validator/any_of_validator'
+require_relative 'schema_validator/all_of_validator'
+require_relative 'schema_validator/one_of_validator'
+require_relative 'schema_validator/nil_validator'
+require_relative 'schema_validator/unspecified_type_validator'
 
 class OpenAPIParser::SchemaValidator
   # validate value by schema
