+++ lib/openapi_parser/schema_validator/all_of_validator.rb
