+++ lib/openapi_parser/schema_validator/any_of_validator.rb
