+++ lib/openapi_parser/schema_validator/array_validator.rb
