+++ lib/openapi_parser/schema_validator/base.rb
