+++ lib/openapi_parser/schema_validator/boolean_validator.rb
