+++ lib/openapi_parser/schema_validator/enumable.rb
