+++ lib/openapi_parser/schema_validator/float_validator.rb
