+++ lib/openapi_parser/schema_validator/integer_validator.rb
