+++ lib/openapi_parser/schema_validator/minimum_maximum.rb
