+++ lib/openapi_parser/schema_validator/nil_validator.rb
