+++ lib/openapi_parser/schema_validator/object_validator.rb
