+++ lib/openapi_parser/schema_validator/one_of_validator.rb
