+++ lib/openapi_parser/schema_validator/options.rb
