+++ lib/openapi_parser/schema_validator/string_validator.rb
