+++ lib/openapi_parser/schema_validator/unspecified_type_validator.rb
